ShopAlgorithms documentation

Main documentation for ShopAlgorithms.jl. This package provides a framework for solving shop scheduling problems. It is a part of a project for my master thesis.

Installation

Currently, this package is not registered in a Julia Packages Registry. To install it, run the following command in the Julia REPL:

import Pkg
Pkg.add(https://github.com/mteplicki/JobShopAlgorithms)

Usage

using ShopAlgorithms
instance = read("path/to/instance", InstanceLoaders.StandardSpecification)
results = Algorithms.bnbcarlier(instance)
println(results)

Documentation

The package consists of the following submodules:

ShopAlgorithms — Module

This module contains the implementation of algorithms for solving job shop scheduling problems. It exports the following modules: Algorithms, ShopInstances, InstanceLoaders, Plotters, Constraints, ShopGraphs.

ShopAlgorithms.Algorithms — Module

Module containing various Job Shop algorithms.

ShopAlgorithms.ShopInstances — Module

This module contains the implementation of various job shop instances, as well as the necessary functions and types to work with them.

ShopAlgorithms.InstanceLoaders — Module

Module for loading and generating random instances of the Shop Problem.

ShopAlgorithms.Plotters — Module

Module for plotting functions and data.

ShopAlgorithms.Constraints — Module

This module contains functions that define constraints for the job shop scheduling problem. Every method in a module should return a function that takes an instance of the problem and returns a boolean value.

ShopAlgorithms.ShopGraphs — Module

This module defines functions and types for working with graphs dedicated for Job Shop algortihms.

Extensions

To extend the package, you can define your own algorithms. To do so, you need to define a function that takes as input an instance and returns a ShopResult object.