Algorithms

Package Algorithms is a main submodule of a project. It provides a set of functions for working with algorithms dedicated to the Job Shop Scheduling Problem.

ShopAlgorithms.Algorithms.algorithm2_two_machines_job_shop — Method

algorithm2_two_machines_job_shop(instance::JobShopInstance; yielding::Bool=false)

Algorithm to solve a J2 | n=k | Cmax problem, with complexity O(r^k), where r = sum(n_i) and k = n.

Arguments

instance::JobShopInstance: An instance of the two machines job shop problem.

yielding::Bool=false: If true, the algorithm will yield after each iteration. This is useful for timeouting the algorithm.

Returns

jobShopSchedule::ShopSchedule: A ShopSchedule object representing the solution to the two machines job shop problem.

ShopAlgorithms.Algorithms.branchandbound — Method

branchandbound(
    instance::JobShopInstance;
    bounding_algorithm::Symbol=:no_pmtn,
    yielding::Bool=false,
    io_print_best_node::Union{IO, Nothing} = nothing,
    heuristic_UB::Bool = false,
)

Branch and Bound algorithm for the Job Shop Scheduling problem J | rcrc | Cmax with recirculation.

Arguments

instance::JobShopInstance: A job shop instance.

bounding_algorithm::Symbol=:no_pmtn: Algorithm used to bound the lower bound of the solution. Possible values are :no_pmtn for 1 | r_j | Lmax and :pmtn for 1 | r_j, pmtn | Lmax. Default value is :no_pmtn.

yielding::Bool=false: If true, the algorithm will yield after each iteration. This is useful for timeouting the algorithm.

io_print_best_node::Union{IO, Nothing} = nothing: If not nothing, the algorithm will print the best node found so far to the given IO.

heuristic_UB::Bool=false: If true, the algorithm will use the Shifting Bottleneck algorithm to find the upper bound of the solution.

Returns

ShopSchedule: A ShopSchedule object representing the solution to the job shop problem.

ShopAlgorithms.Algorithms.branchandbound_carlier — Method

branchandbound_carlier(
    instance::JobShopInstance;
    yielding::Bool = false,
    with_dpc::Bool = false,
    with_priority_queue::Bool = false,
    io_print_best_node::Union{IO, Nothing} = nothing,
    heuristic_UB::Bool = false
)

Branch and Bound algorithm for the Job Shop Scheduling problem J || Cmax with no recirculation.

Arguments

instance::JobShopInstance: A job shop instance.

yielding::Bool=false: If true, the algorithm will yield after each iteration. This is useful for timeouting the algorithm.

with_dpc::Bool=true: If true, the algorithm will use the DPC algorithm to find the longest path in the graph. Otherwise, the algorithm will use the Carlier algorithm.

with_priority_queue::Bool=true: If true, the algorithm will use a priority queue to find the node in Carlier algorithm. Otherwise, it will use a simple stack.

io_print_best_node::Union{IO, Nothing} = nothing: If not nothing, the algorithm will print the best node found so far to the given IO.

heuristic_UB::Bool=false: If true, the algorithm will use the Shifting Bottleneck algorithm to find the upper bound of the solution.

Returns

ShopSchedule: A ShopSchedule object representing the solution to the job shop problem.

ShopAlgorithms.Algorithms.shiftingbottleneck — Method

shiftingbottleneck(
    instance::JobShopInstance;
    suppress_warnings::Bool = false,
    yielding::Bool = false,
    machine_improving::Bool = true
)

Solves the job shop scheduling J || Cmax problem with no job recirculation using the Shifting Bottleneck algorithm. The solution of the problem is not guaranteed to be optimal. Also, not every instance of the problem can be solved using this algorithm.

Arguments

instance::JobShopInstance: An instance of the job shop scheduling problem.

suppress_warnings: If true, warnings will not be printed.

yielding::Bool=false: If true, the algorithm will yield after each iteration. This is useful for timeouting the algorithm.

machine_improving::Bool=true: If true, the algorithm will try to improve the solution by fixing the disjunctive edges for each machine.

Returns

ShopSchedule <: ShopResult: A ShopSchedule object representing the solution to the job shop problem.

ShopError <: ShopResult: A ShopError object representing the error that occured during the execution of the algorithm.

ShopAlgorithms.Algorithms.shiftingbottleneckcarlier — Method

shiftingbottleneckcarlier(
    instance::JobShopInstance; 
    yielding::Bool = false, 
    with_priority_queue::Bool = true,
    with_dpc::Bool = true, 
    carlier_timeout::Union{Nothing,Float64} = nothing, 
    machine_improving::Bool = true,
    carlier_depth::Int64=typemax(Int64)
    )

Solves the job shop scheduling J || Cmax problem with recirculation allowed using the Shifting Bottleneck algorithm with Carlier algorithm. The solution of the problem is not guaranteed to be optimal.

Arguments

instance::JobShopInstance: An instance of the job shop scheduling problem.

yielding::Bool=false: If true, the algorithm will yield after each iteration. This is useful for timeouting the algorithm.

with_dpc::Bool=true: If true, the algorithm will use the Carlier DPC algorithm to find the longest path in the graph. Otherwise, the algorithm will use the Carlier algorithm.

with_priority_queue::Bool=true: If true, the algorithm will use a priority queue to find the longest path in Carlier algorithm. Otherwise, it will use a simple stack.

carlier_timeout::Union{Nothing,Float64}=nothing: If not nothing, the inner Carlier algorithm will be timeouted after carlier_timeout seconds.

machine_improving::Bool=true: If true, the algorithm will try to improve the solution by fixing the disjunctive edges for each machine.

carlier_depth::Int64=typemax(Int64): If not typemax(Int64), the inner Carlier algorithm finding artificial paths will be limited to carlier_depth depth from the real path.

Returns

ShopSchedule <: ShopResult: A ShopSchedule object representing the solution to the job shop problem. The solution is not guaranteed to be optimal.

ShopAlgorithms.Algorithms.two_jobs_job_shop — Method

two_jobs_job_shop(instance::JobShopInstance; yielding::Bool=false)

Solves the two jobs job shop $J | n=2 | Cmax$ problem for a given instance of JobShopInstance with recirculation and machine repetition allowed. Complexity: O(r log r), where r = sum(n_i) is the number of operations.

Arguments

instance::JobShopInstance: An instance of the job shop problem.

yielding::Bool=false: If true, the algorithm will yield after each iteration. This is useful for timeouting the algorithm.

Returns

ShopSchedule: A ShopSchedule object representing the solution to the job shop problem.

ShopAlgorithms.Algorithms.two_machines_job_shop — Method

two_machines_job_shop(instance::JobShopInstance; yielding::Bool=false)

Solves the two-machine job shop problem J2 | p_ij = 1 | Lmax for the given instance. Complexity is O(r), where r = sum(n_i) is the number of operations.

Arguments

instance::JobShopInstance: An instance of the two-machine job shop problem.

yielding::Bool=false: If true, the algorithm will yield after each iteration. This is useful for timeouting the algorithm.

Returns

An optimal solution to the two-machine job shop problem.