Constraints

Package Constrains provides a set of functions for working with constraints dedicated to the Job Shop Scheduling Problem.

ShopAlgorithms.Constraints.job_equals — Method

job_equals(to::Int)

Returns a function that takes an instance and returns a boolean indicating whether the instance's number of jobs is equal to the given integer to.

Arguments

to::Int: The integer to compare the instance's number of jobs.

Returns

A function that takes an instance and returns a boolean indicating whether the instance's number of jobs is equal to the given integer to.

ShopAlgorithms.Constraints.job_lower_limit — Method

job_lower_limit(limit::Int)

Check if the number of jobs in the instance is greater than or equal to the given limit.

Arguments

limit::Int: The lower limit to check against the number of jobs in the instance.

Returns

A function that takes an instance and returns a boolean value: true if the number of jobs in the instance is greater than or equal to the given limit, false otherwise.

ShopAlgorithms.Constraints.job_recirculation — Method

machine_repetition()

Returns a function indicating whether any machine is repeated in the instance.

Returns

A function that takes an instance and returns a boolean value: true if any machine is repeated in the instance, false otherwise.

ShopAlgorithms.Constraints.job_upper_limit — Method

job_upper_limit(limit::Int)

Check if the number of jobs in the instance is lower than or equal to the given limit.

Arguments

limit::Int: The upper limit to check against the number of jobs in the instance.

Returns

A function that takes an instance and returns a boolean value: true if the number of jobs in the instance is less than or equal to the given limit, false otherwise.

ShopAlgorithms.Constraints.machine_equals — Method

machine_equals(to::Int)

Returns a function that takes an instance and returns a boolean indicating whether the instance's m field is equal to the given integer to.

Arguments

to::Int: The integer to compare the instance's m field to.

Returns

A function that takes an instance and returns a boolean indicating whether the instance's m field is equal to the given integer to.

ShopAlgorithms.Constraints.machine_lower_limit — Method

machine_lower_limit(limit::Int)

Checks if the lower limit of the machine is greater than or equal to the given limit.

Arguments

limit::Int: the limit to check against the lower limit of the machine.

Returns

function that takes an instance of the problem and returns a boolean value: true if the lower limit of the machine is greater than or equal to the given limit, false otherwise.

ShopAlgorithms.Constraints.machine_repetition — Method

machine_repetition()

Returns a function indicating whether any jobs has to be processed in the same machine more than once in a row.

Returns

A function that takes an instance and returns a boolean value: true if any jobs has to be processed in the same machine more than once in a row, false otherwise.

ShopAlgorithms.Constraints.machine_upper_limit — Method

machine_upper_limit(limit::Int)

Checks if the upper limit of the machine is less than or equal to the given limit.

Arguments

limit::Int: The maximum limit allowed for the machine.

Returns

a function that takes an instance of the problem and returns a boolean value: true if the upper limit of the machine is less than or equal to the given limit, false otherwise.

ShopAlgorithms.Constraints.processing_time_lower_limit — Method

processing_time_lower_limit(limit::Int)

Returns a funtion indicating whether all processing times in the instance are greater than or equal to the given limit.

Arguments

limit::Int: The lower limit for processing times.

Returns

A function that takes an instance and returns a boolean value: true if all processing times in the instance are greater than or equal to the given limit, false otherwise.

ShopAlgorithms.Constraints.processing_time_upper_limit — Method

processing_time_upper_limit(limit::Int)

Returns a function indicating whether all processing times in the instance are less than or equal to the given limit.

Arguments

limit::Int: The upper limit for processing times.

Returns

A function that takes an instance and returns a boolean value: true if all processing times in the instance are less than or equal to the given limit, false otherwise.