Instance Loaders

The InstanceLoaders module provides functions to load instances from files. It also provides functions to generate random instances.

Loading instances from files

The Base.read function can be used to load an instance from a file. It takes as input the path to the file and returns an instance. There are two types of specification: StandardSpecification and TaillardSpecification.

ShopAlgorithms.InstanceLoaders.StandardSpecification — Type

StandardSpecification <: JobShopFileSpecification

A struct that represents a standard job shop instance specification. Standard instances are specified in the following format:

n m
μ_11 p_11 μ_12 p_12 ... μ_1n_1 p_1n_1
μ_21 p_21 μ_22 p_22 ... μ_2n_2 p_2n_2
...
μ_m1 p_m1 μ_m2 p_m2 ... μ_mn_m p_mn_m

where n is the number of jobs, m is the number of machines, p_ij is the processing time of job i on machine j, and μ_ij is the machine number of the jth operation of job i, starting from 0.

Fields

instance::JobShopInstance: A job shop instance.

Examples

julia> instance = JobShopInstance(3, 2, [2, 2, 2], [[1, 2], [2, 1], [1, 2]], [[1, 2], [2, 1], [1, 2]])
JobShopInstance(3, 2, [2, 2, 2], [[1, 2], [2, 1], [1, 2]], [[1, 2], [2, 1], [1, 2]])
julia> specification = StandardSpecification(instance)
StandardSpecification(JobShopInstance(3, 2, [2, 2, 2], [[1, 2], [2, 1], [1, 2]], [[1, 2], [2, 1], [1, 2]]))
julia> write("test.txt", specification);
julia> read("test.txt", StandardSpecification)
StandardSpecification(JobShopInstance(3, 2, [2, 2, 2], [[1, 2], [2, 1], [1, 2]], [[1, 2], [2, 1], [1, 2]]))

ShopAlgorithms.InstanceLoaders.TaillardSpecification — Type

TaillardSpecification <: JobShopFileSpecification

A struct that represents a Taillard instance specification for the Job Shop Problem. Taillard instances are specified in the following format:

n m
p_11 p_12 ... p_1n_1
p_21 p_22 ... p_2n_2
...
p_m1 p_m2 ... p_mn_m
μ_11 μ_12 ... μ_1n_1
μ_21 μ_22 ... μ_2n_2
...
μ_m1 μ_m2 ... μ_mn_m

where n is the number of jobs, m is the number of machines, p_ij is the processing time of job i on machine j, and μ_ij is the machine number of the jth operation of job i.

Fields

instance::JobShopInstance: A JobShopInstance object representing the instance.

Examples

julia> instance = JobShopInstance(3, 2, [2, 2, 2], [[1, 2], [2, 1], [1, 2]], [[1, 2], [2, 1], [1, 2]])
JobShopInstance(3, 2, [2, 2, 2], [[1, 2], [2, 1], [1, 2]], [[1, 2], [2, 1], [1, 2]])
julia> specification = TaillardSpecification(instance)
TaillardSpecification(JobShopInstance(3, 2, [2, 2, 2], [[1, 2], [2, 1], [1, 2]], [[1, 2], [2, 1], [1, 2]]))
julia> write("test.txt", specification)
julia> read("test.txt", TaillardSpecification)
TaillardSpecification(JobShopInstance(3, 2, [2, 2, 2], [[1, 2], [2, 1], [1, 2]], [[1, 2], [2, 1], [1, 2]]))

# Reading a instance
Base.read(data::IO, ::Type{T}) where {T <: JobShopFileSpecification} 

Saving instances to files

The Base.write function can be used to save an instance to a file. It takes as input the path to the file and the instance.

# Writing a instance
Base.write(data::IO, specification::T) where {T <: JobShopFileSpecification}

Generating random instances

The InstanceLoaders module provides functions to generate random instances. The random_instance_generator function can be used to generate a random instance.

ShopAlgorithms.InstanceLoaders.random_instance_generator — Function

random_instance_generator(n::Int64, m::Int64; n_i::Union{Vector{Int}, Nothing}=nothing, pMin::Int=1, pMax::Int=10, rng=default_rng(), job_recirculation::Bool=false, machine_repetition::Bool=false)::JobShopInstance

This function generates a random instance of the Job Shop Scheduling problem.

Arguments

n::Int64: number of jobs.

m::Int64: number of machines.

n_i::Union{Vector{Int}, Nothing}=nothing: number of operations for each job. If nothing, all jobs have the same number m of operations.

pMin::Int=1: minimum processing time for an operation.

pMax::Int=10: maximum processing time for an operation.

rng=default_rng(): random number generator.

job_recirculation::Bool=false: if true, a job can be processed in the same machine more than once.

machine_repetition::Bool=false: if true, an operation can be processed in the same machine more than once in a row. If false, an operation cannot be processed in the same machine as the previous operation.

Returns

JobShopInstance: a random instance of the Job Shop Scheduling problem.