Plotters

The Plotters module provides functions to plot the results of the algorithms.

Plotting the results of an algorithm

The gantt_chart function can be used to plot the gantt chart of an algorithm. It takes as input the algorithm results and the instance. The algorithm results can be obtained by via proper function from module Algortihms.

ShopAlgorithms.Plotters.gantt_chart — Function

gantt_chart(solution::ShopSchedule)

Create a Gantt chart for a given ShopSchedule solution.

Arguments

solution::ShopSchedule: A ShopSchedule object representing a solution to a job shop scheduling problem.

Returns

A plot of the solution using a Gantt chart.

The plot_geometric_approach function can be used to plot the geometric approach of an two jobs problem.

ShopAlgorithms.Plotters.plot_geometric_approach — Function

plot_geometric_approach(solution::ShopSchedule; title::Union{Nothing,String}=nothing, width=800, height=800, aspectmode="auto")

This function takes a ShopSchedule object as input and returns a plot of the solution using a geometric approach, when n == 2. Plot is generated using PlotlyJS.

Arguments

solution::ShopSchedule: A ShopSchedule object representing the solution to a job shop scheduling problem.

title::Union{Nothing,String}=nothing: Title of the plot. If nothing, the title will be the name of the instance.

width=800: Width of the plot.

height=800: Height of the plot.

aspectmode="auto": Aspect mode of the plot.

Returns

A plot of the solution using a geometric approach.

Examples

julia> instance = random_instance_generator(2,5);
julia> solution = two_jobs_job_shop(instance);
julia> plot_geometric_approach(solution)