ShopGraphs

Package ShopGraphs provides a set of functions for working with graphs dedicated to the Job Shop Scheduling Problem.

ShopAlgorithms.ShopGraphs.DiWeightedEdge — Type

mutable struct DiWeightedEdge{T<:Integer, U<:Real} <: AbstractEdge{T}

    DiWeightedEdge is a mutable struct that represents a weighted edge in a graph. It contains the source vertex, destination vertex, and weight of the edge.

Arguments

src::T: The source vertex of the edge.

dst::T: The destination vertex of the edge.

weight::U: The weight of the edge.

Type parameters

T<:Integer: The type of the vertices.

U<:Real: The type of the weight.

ShopAlgorithms.ShopGraphs.DisjunctiveWeightedGraph — Type

mutable struct DisjunctiveWeightedGraph{T<:Integer, U<:Real} <: AbstractGraph{T}

DisjunctiveWeightedGraph is a mutable struct that represents a disjunctive weighted graph. It contains two fields:

conjunctiveGraph: a SimpleDirectedWeightedGraphAdj{T,U} representing the conjunctive graph.

selection: a SimpleDirectedWeightedGraphAdj{T,U} representing the selection of disjunctive graph.

Arguments

T::Type{<:Integer}: The integer type of the vertices.

U::Type{<:Real}: The real type of the weights.

Fields

conjunctiveGraph::SimpleDirectedWeightedGraphAdj{T,U}: The conjunctive graph.

selection::SimpleDirectedWeightedGraphAdj{T,U}: The selection of disjunctive graph.

Constructors

DisjunctiveWeightedGraph{T,U}(conjunctiveGraph::SimpleDirectedWeightedGraphAdj{T,U}, selection::SimpleDirectedWeightedGraphAdj{T,U}): Constructs a new DisjunctiveWeightedGraph{T,U} with the given conjunctiveGraph and selection.

Examples

    conjuctiveGraph = SimpleDirectedWeightedGraphAdj{Int,Int}([
        [SimpleDirectedWeightedEdge(1, 2, 1), SimpleDirectedWeightedEdge(1, 3, 1)],
        [SimpleDirectedWeightedEdge(2, 4, 1)],
        [SimpleDirectedWeightedEdge(3, 4, 1)],
        [SimpleDirectedWeightedEdge(4, 5, 1)]
    ])
    selection = SimpleDirectedWeightedGraphAdj{Int,Int}([
        [SimpleDirectedWeightedEdge(1, 2, 1), SimpleDirectedWeightedEdge(1, 3, 1)],
        [SimpleDirectedWeightedEdge(2, 4, 1)],
        [SimpleDirectedWeightedEdge(3, 4, 1)],
        [SimpleDirectedWeightedEdge(4, 5, 1)]
    ])
    disjunctiveWeightedGraph = DisjunctiveWeightedGraph(conjuctiveGraph, selection)

ShopAlgorithms.ShopGraphs.SimpleDiWeightedGraphAdj — Type

SimpleDiWeightedGraphAdj{T<:Integer, U<:Real}

A simple directed weighted graph implementation using an adjacency list representation.

Fields

edges::Vector{Vector{DiWeightedEdge{T,U}}}: A vector of out edges represented as a vector of DiWeightedEdge{T,U}.

edges_transpose::Vector{Vector{DiWeightedEdge{T,U}}}: A vector of in edges represented as a vector of DiWeightedEdge{T,U}.

Constructors

SimpleDiWeightedGraphAdj{T,U}(edges::Vector{Vector{DiWeightedEdge{T,U}}}): Constructs a new SimpleDiWeightedGraphAdj{T,U} object with the given vertices and edges.

ShopAlgorithms.ShopGraphs.dag_paths — Method

dag_paths(graph::AbstractGraph, source::V, type::Symbol) where {V<:Integer, U<:Real}

Returns all paths from the source vertex to all other vertices in a directed acyclic graph (DAG).

Arguments

graph::AbstractGraph: A directed acyclic graph.

source::V: The source vertex.

type::Symbol: The type of path to return. Possible values are :shortest and :longest.

reversed::Bool: Whether to use the reversed graph.

Returns

dist::Vector{V}: A vector of distances from the source vertex to all other vertices.

Throws

ArgumentError: If the type is not :shortest or :longest, or if the graph is not a DAG.

ShopAlgorithms.ShopGraphs.topological_sort_util — Method

topological_sort_util(graph::AbstractGraph, visited::BitVector, pathVis::BitVector, stack::Stack{V}, v::V) where {V<:Integer, U<:Real}

A utility function used in topological sorting of a directed acyclic graph (DAG). This function recursively visits all the vertices adjacent to the given vertex v and adds them to the stack in topological order. It also detects cycles in the graph by keeping track of the visited vertices and the vertices in the current path.

Arguments

graph::AbstractGraph: The directed acyclic graph to be sorted.

visited::BitVector: A bit vector to keep track of the visited vertices.

pathVis::BitVector: A bit vector to keep track of the vertices in the current path.

stack::Stack{V}: A stack to store the vertices in topological order.

v::V: The vertex to be visited.

reversed::Bool: Whether to use the reversed graph.

Returns

nothing

Throws

ArgumentError: If the graph is not a DAG.