ShopInstances

Package to define and work with instances of the Job Shop Scheduling Problem.

ShopAlgorithms.ShopInstances.AbstractShop — Type

AbstractShop

Abstract type representing a shop problem.

ShopAlgorithms.ShopInstances.JobShopInstance — Type

struct JobShopInstance <: AbstractShop

JobShopInstance(
    n::Int64,
    m::Int64,
    n_i::Vector{Int},
    p::Vector{Vector{Int}},
    μ::Vector{Vector{Int}};
    d::Vector{Int}=zeros(Int64, n),
    name::String=""
)

JobShopInstance represents a job shop instance, which is a type of scheduling problem.

Arguments

n::Int64: number of jobs

m::Int64: number of machines

n_i::Vector{Int}: number of operations for each job

p::Vector{Vector{Int}}: processing times for each operation

μ::Vector{Vector{Int}}: machines assigned to each operation

d::Vector{Int} = zeros(Int64, n): due dates for each job

name::String = "": name of the instance

Fields

n: an integer representing the number of jobs

m: an integer representing the number of machines

n_i: a vector of integers representing the number of operations for each job

p: a vector of vectors of integers representing the processing times for each operation

μ: a vector of vectors of integers representing the machines assigned to each operation

d: a vector of integers representing the due dates for each job

name: a string representing the name of the instance

ShopAlgorithms.ShopInstances.ObjectiveFunction — Type

Enum representing different objective functions for shop scheduling problems.

ShopAlgorithms.ShopInstances.ShopError — Type

ShopError <: ShopResult

ShopError(
    instance::AbstractShop,
    error::String,
    objectiveFunction::ObjectiveFunction;
    algorithm::String="",
    metadata::Dict{String, Any}=Dict{String, Any}(),
    date::DateTime=now()
)

A struct representing an error result of a shop scheduling algorithm.

Arguments

instance::AbstractShop: The instance of the shop problem.

error::String: The error message.

objectiveFunction::ObjectiveFunction: The objective function used in the algorithm.

`algorithm::String="": The name of the algorithm used.

metadata::Dict{String, Any}=Dict{String, Any}(): Additional metadata about the result.

date::DateTime=now(): The date and time when the result was obtained.

Fields

instance::JobShopInstance: The instance of the shop problem.

error::String: The error message.

objectiveFunction::ObjectiveFunction: The objective function used in the algorithm.

algorithm::String: The name of the algorithm used.

date::DateTime: The date and time when the result was obtained.

metadata::Dict{String, Any}: Additional metadata about the result.

ShopAlgorithms.ShopInstances.check_feasability — Method

check_feasability(schedule::ShopSchedule)

Check the feasability of a given schedule for a job-shop problem instance.

Arguments

schedule::ShopSchedule: A schedule for a job-shop problem instance.

Returns

true if the schedule is feasible, false otherwise.

ShopAlgorithms.ShopInstances.dataframe_to_schedules — Method

dataframe_to_schedules(df::DataFrame)::Vector{ShopSchedule}

Converts a DataFrame to a Vector of ShopSchedule objects. The DataFrame should have the following columns:

solution_id: unique identifier for each solution

name: name of the instance

m: number of machines

n: number of jobs

algorithm: name of the algorithm used to generate the solution

microruns: number of microruns used by the algorithm

timeSeconds: time in seconds used by the algorithm

memoryBytes: memory used by the algorithm

job: job number

operation: operation number

endtime: end time of the operation

processing_time: processing time of the operation

machine: machine number where the operation was executed

d: due date of the job

Returns a Vector of ShopSchedule objects, each representing a solution in the DataFrame.